[FIX] Move core/tuple_utility.hpp to utility/tuple_utility.hpp
- moves tuple_utility.hpp from [core] to [utility]
- adjusts all header files that include tuple_utility.hpp
- moves related test and snippets accordingly
- adjusts CMakeFiles.txt accordingly
- creates new deprecated tuple_utility.hpp header file
- removes tuple_utility.hpp from core/all.hpp
- adjusts doxygen `\ingroup` tag and snippets path
