+++ include/seqan3/core/all.hpp
@@ -26,7 +26,6 @@
 #include <seqan3/core/platform.hpp>
 #include <seqan3/core/pod_tuple.hpp>
 #include <seqan3/core/range/type_traits.hpp>
-#include <seqan3/core/tuple_utility.hpp>
 
 /*!\defgroup core Core
  * \brief Provides core functionality used by multiple modules.
