+++ test/unit/core/CMakeLists.txt
@@ -4,4 +4,3 @@ seqan3_test (add_enum_bitwise_operators_test.cpp)
 seqan3_test (bit_manipulation_test.cpp)
 seqan3_test (debug_stream_test.cpp)
 seqan3_test (pod_tuple_test.cpp)
-seqan3_test (tuple_utility_test.cpp)
