+++ test/unit/utility/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectories()
 
 seqan3_test (math_test.cpp)
+seqan3_test (tuple_utility_test.cpp)
